feat: Enhance bike tracking with weather, refactor sensor data handling
This commit enhances the bike tracking feature by adding real-time weather information and refactors sensor data handling for improved efficiency and clarity.

Key changes:

- Integrated weather data into `BikeRideInfo` using `BikeWeatherInfo`.
- Implemented weather fetching based on the initial location received from sensors.
- Updated `BikeViewModel` to merge weather data into the UI state.
- Enhanced sensor data processing by throttling updates to 200ms and performing computations on a background thread.
- Refactored `WindDirectionDialWithSpeed` to use `Int` for wind degree and `Double` for wind speed.
- Updated location permission handling in `MainActivity`.
- Added `ACCESS_BACKGROUND_LOCATION` permission to the manifest.

+++ core/model/src/main/java/com/ylabz/basepro/core/model/weather/BikeWeatherInfo.kt
@@ -1,7 +1,7 @@
 package com.ylabz.basepro.core.model.weather
 
 data class BikeWeatherInfo(
-    val windDegree: Float,
-    val windSpeed: Float,
+    val windDegree: Int,
+    val windSpeed: Double,
     val conditionText: String
 )
